Refactor: Lights reference coords, not cells
Thinking about how I'd make changes to a grid (e.g. a partially
completed one) I realised that the previous structure was too
denormalized, and would be a nightmare to update in a functional
structure.  So instead of Lights containing a list of [GridCell],
they now contain just [Coord].  This means:

 - Crow now needs a function to look up Lights at a Coord
 - Light can't `stringify` its contents (as it doesn't know them,
   without an explicit reference to Crow)
 - GridCell not needed, but instead the Grid is zipped with its
   coordinates as required (with the lovely `zipWith zip`)