Progress towards getLightsForAnswer
We're going to want to create Clues soon, which means we'll need a way
to determine which Lights they span.

We could pass in tuples like `(5, Across)`, but this means that the
setter will have to calculate the light numbers, even though the code is
smart enough to do this for us!

Instead, it seems much more pleasant to pass in the answer itself, like
["pocket", "watch"] and find out which lights contain those words.  Some
setters (like Araucaria RIP) like to spread a single word over multiple
lights, e.g. "ICEBOX (6)" might actually be spread over two lights
("ICE" and "BOX").  And I suspect setters might do clever things like
spread an answer "TO THE MAX" over lights containing ("TOT", "HEM",
"AX")

So the algorithm is to simply concatenate the answer, and then check
whether any combination of lights add up to it.  In case some random
combination of shorter words matches a word we wanted to be mapped to a
longer light, we prefer longer matches.

(Current function simply matches on strings, but simply needs to be
plumbed in to Lights via the mapping function 'f')

(Thanks to lyxia for some code suggestions, including using
`stripPrefix` instead of a guard + dropPrefix, and replacing an explicit
accumulator with the `fmap` expression.)

Also tidied up imports.