Add barebones snap frontend
As a first web-project with Haskell, I found this a little confusing and
undocumented, especially using Snap with a Stack project (apparently
most devs are using plain Cabal for this) so I'll document the steps
here.

1) add ', snap' to build-depends in crow.cabal
2) stack build
3) mkdir snap/; cd snap
4) stack exec snap init barebones
5) cp src/Main.hs ../app/Main.hs
6) cp -r logs/ ..
7) vim snap.cabal crow.cabal  # see detail below
8) stack build
9) stack exec crow-exe -- --port 5000
10) visit http://localhost:5000/

Steps 1-2 are simply to get Snap installed within the Stack environment.

Steps 3-4 create a barebones snap project. Because the `snap` executable
is designed to start a new project, not edit an existing one, we have to
do it in a new directory.

Step 5: copy the barebones file.  In addition, I imported the Crow
library, just because (but had to do it qualified, as the 'dir' symbol
clashes)

Step 6: the snap server will complain if there's no logs/ directory.  I
copied the empty access.log, so that it can sit in my git tree, to
ensure the directory exists.

Step 7: see the diff for details.  Basically, I copied the dependencies
from snap.cabal to the section for 'crow-exe' (e.g. the executable that
Stack had created for me).  I removed the min/max versions that were in
the snap.cabal, on the basis that stack is supposed to free me from
worrying about those.
NB: I also got rid of the base 'snap' line that I added in Step 1, as
that was just to bootstrap the installation.

Step 8-9: build the executable and run it.  I'm running on a specific
port because that's tunnelled from my Vagrant dev environment.  I have
to add the '--' to make sure that it's 'crow-exe' that gets the argument
rather than 'stack' itself.